Export getRange to helper funciton
